Fix #336 Move decompress action to main process
